Experiment-8/config.py
from dotenv import load_dotenv
import os

load_dotenv()

# Twitter API Credentials
TWITTER_API_KEY = os.getenv("TWITTER_API_KEY")
TWITTER_API_SECRET = os.getenv("TWITTER_API_SECRET")
TWITTER_ACCESS_TOKEN = os.getenv("TWITTER_ACCESS_TOKEN")
TWITTER_ACCESS_TOKEN_SECRET = os.getenv("TWITTER_ACCESS_TOKEN_SECRET")

# Sentiment Thresholds
SENTIMENT_THRESHOLDS = {
    'positive': 0.1,
    'negative': -0.1,
}